envoie d'un user associé à une promo fait

+++ BackEnd/MMIPlatform/src/main/java/MMIPlatform/MMIPlatform/Repositories/PromoRepository.java
@@ -3,6 +3,6 @@ package MMIPlatform.MMIPlatform.Repositories;
 import MMIPlatform.MMIPlatform.Models.Promo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PromoRepository extends JpaRepository<Promo, Long> {
+public interface PromoRepository extends JpaRepository<Promo, Integer> {
 
 }
